fix : 引入jpa，替换mp，为适配数据分片做打算

+++ big-market-domain/src/main/java/app/xlog/ggbond/raffle/model/AwardBO.java
@@ -19,8 +19,8 @@ public class AwardBO {
     private Integer awardId;
     private String awardTitle;
     private String awardSubtitle;
-    private Integer awardCount;
-    private float awardRate;
+    private Long awardCount;
+    private Double awardRate;
     private Integer awardSort;
     private String rules;
